Move MicrocontrollerFactory to the orchestrator module to break a circular dependency
