+++ ts/libs/core/src/lib/orchestrator/public/error.ts
@@ -145,3 +145,12 @@ export class ObjectKindsAlreadyWatchedError extends Error {
     }
 
 }
+
+
+export class MicrocontrollerFactoryNotRegisteredError extends Error {
+
+    constructor(public spec: any) {
+        super('No `FactoryFn` was registered for the spec.');
+    }
+
+}
