+++ ts/libs/core/src/lib/orchestrator/public/impl/default-microcontroller-factory.ts
@@ -1,7 +1,8 @@
 import { PolarisMicrocontrollerFactoryMetadata } from '../../../transformation/internal/polaris-microcontroller-factory-metadata';
 import { PolarisMetadataUtils } from '../../../transformation/internal/reflect-metadata-utils';
 import { Constructor, FactoryFn } from '../../../util';
-import { MicrocontrollerFactory, MicrocontrollerFactoryNotRegisteredError } from '../microcontroller-factory';
+import { MicrocontrollerFactoryNotRegisteredError } from '../error';
+import { MicrocontrollerFactory } from '../microcontroller-factory';
 
 export class DefaultMicrocontrollerFactory<S, C> implements MicrocontrollerFactory<S, C> {
