+++ ts/libs/core/src/lib/orchestrator/public/impl/index.ts
@@ -1,1 +1,2 @@
+export * from './default-microcontroller-factory';
 export * from './default-watch-manager';
