+++ ts/libs/core/src/lib/orchestrator/public/microcontroller-factory.ts
@@ -1,4 +1,4 @@
-import { Constructor, FactoryFn } from '../../../util';
+import { Constructor, FactoryFn } from '../../util';
 
 /**
  * Used to instantiate a microcontroller for a particular Spec instance, e.g., a `ServiceLevelObjective`
