+++ ts/libs/core/src/lib/runtime/public/impl/index.ts
@@ -1,2 +1,1 @@
-export * from './default-microcontroller-factory';
 export * from './polaris-runtime.base';
