+++ ts/libs/core/src/lib/runtime/public/index.ts
@@ -1,2 +1,1 @@
-export * from './microcontroller-factory';
 export * from './polaris-runtime';
