--- ts/libs/core/src/lib/runtime/public/microcontroller-factory/errors.ts
@@ -1,8 +1,0 @@
-
-export class MicrocontrollerFactoryNotRegisteredError extends Error {
-
-    constructor(public spec: any) {
-        super('No `FactoryFn` was registered for the spec.');
-    }
-
-}
