--- ts/libs/core/src/lib/runtime/public/microcontroller-factory/index.ts
@@ -1,2 +1,0 @@
-export * from './errors';
-export * from './microcontroller-factory';
