--- ts/libs/core/src/lib/runtime/public/polaris-runtime/index.ts
@@ -1,1 +1,0 @@
-export * from './polaris-runtime';
